Allow artd and profman to access package installing tmp folder
This change is needed for refactoring the installation flow and move dexopt to be before the installation commit starts.

When we try to do dexopt at the temporary folder of the app installing, we encounter an error, because artd has no permissioin to access the tmp folder. Also, art uses profman during dexopt, and profman has also no permission.

Bug: 321139675
Test: manual
Change-Id: I6e8d3fe805d28ce8ffaf8768d4aa9428edaf81a3