sepolicy: add rules for bionic.linker.16kb.app_compat.enabled
This property serves as the global toggle for 16kb app compat.

It must be world readable since the linker is executed in the context
of the process requesting the dynamic linking.

Currently, only init and root shell are allowed to set the property.
This may be modified later to also allow toggling from the settings
app.

Bug: 339709616
Test: m
Change-Id: Ie0fd4132477460d88c3cf2f09aa59084f54cfea7
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>